requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
python-dotenv==1.0.1
Pillow==10.4.0
PyMuPDF==1.24.14
yookassa>=3.0.0,<4.0.0


